requirements.txt
pandas==2.2.3
psycopg2_binary==2.9.10
python-dotenv==1.0.1
